+++ src/euroeval/dataset_configs/__init__.py
@@ -5,6 +5,7 @@ from ..languages import get_all_languages
 from ..tasks import SPEED
 from ..utils import load_custom_datasets_module
 from .bulgarian import *  # noqa: F403
+from .croatian import *  # noqa: F403
 from .czech import *  # noqa: F403
 from .danish import *  # noqa: F403
 from .dutch import *  # noqa: F403
@@ -24,6 +25,7 @@ from .polish import *  # noqa: F403
 from .portuguese import *  # noqa: F403
 from .serbian import *  # noqa: F403
 from .slovak import *  # noqa: F403
+from .slovenian import *  # noqa: F403
 from .spanish import *  # noqa: F403
 from .swedish import *  # noqa: F403
 from .ukrainian import *  # noqa: F403