+++ src/euroeval/prompt_templates/sentiment_classification.py
@@ -5,6 +5,7 @@ import typing as t
 from ..data_models import PromptConfig
 from ..languages import (
     BULGARIAN,
+    CROATIAN,
     CZECH,
     DANISH,
     DUTCH,
@@ -26,6 +27,7 @@ from ..languages import (
     PORTUGUESE,
     SERBIAN,
     SLOVAK,
+    SLOVENIAN,
     SPANISH,
     SWEDISH,
     UKRAINIAN,
@@ -54,6 +56,19 @@ SENT_TEMPLATES: dict["Language", PromptConfig] = {
         default_prompt_template="Dokument: {text}\nSentiment: {label}",
         default_instruction_prompt="Dokument: {text}\n\nKlassificer sentimentet i "
         "dokumentet. Svar kun med {labels_str}, og intet andet.",
+    ),
+    CROATIAN: PromptConfig(
+        default_prompt_label_mapping=dict(
+            positive="pozitivno", neutral="neutralno", negative="negativno"
+        ),
+        default_prompt_prefix=(
+            "Slijede dokumenti i njihova osjetila, koja mogu biti {labels_str}."
+        ),
+        default_prompt_template=("Dokument: {text}\nOsjetilo: {label}"),
+        default_instruction_prompt=(
+            "Dokument: {text}\n\nKlasificirajte osjećaj u dokumentu. "
+            "Odgovorite samo s {labels_str}, i ništa drugo."
+        ),
     ),
     CZECH: PromptConfig(
         default_prompt_label_mapping=dict(
@@ -259,6 +274,20 @@ SENT_TEMPLATES: dict["Language", PromptConfig] = {
         default_prompt_template="Dokument: {text}\nSentiment: {label}",
         default_instruction_prompt="Dokument: {text}\n\nKlasifikujte pocit v "
         "dokumente. Odpovedzte so {labels_str}, a nič iné.",
+    ),
+    SLOVENIAN: PromptConfig(
+        default_prompt_label_mapping=dict(
+            positive="pozitivno", neutral="nevtralno", negative="negativno"
+        ),
+        default_prompt_prefix=(
+            "Spodaj so dokumenti in njihov sentiment, ki je lahko {labels_str}."
+        ),
+        default_prompt_template=("Dokument: {text}\nSentiment: {label}"),
+        default_instruction_prompt=(
+            "Dokument: {text}\n\n"
+            "Klasificirajte sentiment v dokumentu. "
+            "Odgovorite z {labels_str}, in nič drugega."
+        ),
     ),
     SERBIAN: PromptConfig(
         default_prompt_label_mapping=dict(