+++ src/scripts/create_multi_wiki_qa.py
@@ -48,6 +48,8 @@ LANGUAGES = [
     "el",
     "bg",
     "sr",
+    "sl",
+    "hr",
 ]