+++ tests/test_benchmarker.py
@@ -42,8 +42,8 @@ def test_benchmark_results_is_a_list(benchmarker: Benchmarker) -> None:
     assert isinstance(benchmarker.benchmark_results, list)
 
 
-@pytest.mark.flaky(reruns=3, reruns_delay=5)
 @pytest.mark.depends(on=["tests/test_model_loading.py::test_load_non_generative_model"])
+@pytest.mark.flaky(reruns=3, reruns_delay=5)
 def test_benchmark_encoder(
     benchmarker: Benchmarker, task: Task, language: Language, encoder_model_id: str
 ) -> None:
@@ -66,6 +66,7 @@ def test_benchmark_encoder(
     condition=not torch.cuda.is_available(), reason="CUDA is not available."
 )
 @pytest.mark.depends(on=["tests/test_model_loading.py::test_load_generative_model"])
+@pytest.mark.flaky(reruns=3, reruns_delay=5)
 def test_benchmark_generative(
     benchmarker: Benchmarker, task: Task, language: Language, generative_model_id: str
 ) -> None:
@@ -81,6 +82,7 @@ def test_benchmark_generative(
     condition=not torch.cuda.is_available(), reason="CUDA is not available."
 )
 @pytest.mark.depends(on=["tests/test_model_loading.py::test_load_generative_model"])
+@pytest.mark.flaky(reruns=3, reruns_delay=5)
 def test_benchmark_generative_adapter(
     benchmarker: Benchmarker,
     task: Task,
@@ -99,7 +101,6 @@ def test_benchmark_generative_adapter(
     condition=os.getenv("OPENAI_API_KEY") is None,
     reason="OpenAI API key is not available.",
 )
-@pytest.mark.flaky(reruns=3, reruns_delay=5)
 def test_benchmark_openai(
     benchmarker: Benchmarker, task: Task, language: Language, openai_model_id: str
 ) -> None:
@@ -114,7 +115,6 @@ def test_benchmark_openai(
 @pytest.mark.skipif(
     condition=os.system("uv run ollama -v") != 0, reason="Ollama is not available."
 )
-@pytest.mark.flaky(reruns=3, reruns_delay=5)
 def test_benchmark_ollama(
     benchmarker: Benchmarker, task: Task, language: Language, ollama_model_id: str
 ) -> None: